Add the rest of the curve stuff
* The base curve can now be edited like the nose curve.
* Added an option to change the number of segments that the fairing is
comprised (both base and nose).
* Reshuffled the GUI elements a bit to make handling a bit easier.
* Reverted the fairing side part name for compatibility with various
mods that use MM to clone and edit the baseline fairing part.

+++ Source/ProceduralFairings/FairingBase.cs
@@ -773,8 +773,8 @@ namespace Keramzit
                 minBaseConeAngle = sf.minBaseConeAngle;
                 baseConeShape = sf.baseConeShape;
                 noseConeShape = sf.noseConeShape;
-                baseConeSegments = sf.baseConeSegments;
-                noseConeSegments = sf.noseConeSegments;
+                baseConeSegments = (int) sf.baseConeSegments;
+                noseConeSegments = (int) sf.noseConeSegments;
                 mappingScale = sf.mappingScale;
                 stripMapping = sf.stripMapping;
                 horMapping = sf.horMapping;
